# backend/internal/httpapi/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Crypto Scanner API
  version: 1.0.0
  description: |
    HTTP API for service health, cryptocurrency market analysis, and candle history.

    Live candles use the same-origin WebSocket endpoint `/api/v1/live/candles`.
    Clients exchange Telegram Mini App init data for a session token with
    `POST /api/v1/auth/session` and send `Authorization: Bearer <token>` to every
    other operation. A request rejected with `401` needs a new session.

    The client first sends `LiveCandleClientMessage` with type `authenticate` and
    the session token; the token is never placed in the URL. After `authenticated`, the
    connection accepts dynamic `subscribe` and `unsubscribe` messages and emits
    `LiveCandleServerMessage` snapshots, updates, freshness changes, acknowledgements,
    and bounded errors. One connection may switch subscriptions without reconnecting.
servers:
  - url: /
...
paths:
...
  /api/v1/auth/session:
    post:
      operationId: createSession
      summary: Exchange Telegram init data for a session token
      tags: [Authentication]
      security: [{ telegramInitData: [] }]
      parameters:
        - $ref: "#/components/parameters/RequestID"
      responses:
        "201":
          description: Session created.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content: { application/json: { schema: { $ref: "#/components/schemas/Session" } } }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/AccessDenied" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      operationId: deleteSession
      summary: Revoke the current session
      tags: [Authentication]
      security: [{ sessionToken: [] }]
      parameters:
        - $ref: "#/components/parameters/RequestID"
      responses:
        "204": { description: Session revoked. }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/AccessDenied" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/analysis/instruments/{symbol}:
    post:
      operationId: analyzeInstrument
      summary: Analyze one active instrument
      tags: [Analysis]
      security:
        - sessionToken: []
      parameters:
        - $ref: "#/components/parameters/RequestID"
        - $ref: "#/components/parameters/Symbol"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstrumentAnalysisRequest"
      responses:
        "200":
          description: Instrument analysis result.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstrumentAnalysisResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/AccessDenied"
        "404":
          $ref: "#/components/responses/SymbolNotFound"
        "409":
          $ref: "#/components/responses/InsufficientData"
        "422":
          $ref: "#/components/responses/UnprocessableAnalysis"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/AnalysisUnavailable"
  /api/v1/analysis/market:
    post:
      operationId: analyzeMarket
      summary: Analyze active instruments in the market
      tags: [Analysis]
      security:
        - sessionToken: []
      parameters:
        - $ref: "#/components/parameters/RequestID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketAnalysisRequest"
      responses:
        "200":
          description: Market analysis result.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketAnalysisResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/AccessDenied"
        "422":
          $ref: "#/components/responses/UnprocessableAnalysis"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/AnalysisUnavailable"
  /api/v1/instruments/{symbol}/candles:
    get:
      operationId: listInstrumentCandles
      summary: List a chronological page of closed candles
      tags: [Market Data]
      security:
        - sessionToken: []
      parameters:
        - $ref: "#/components/parameters/RequestID"
        - $ref: "#/components/parameters/Symbol"
        - name: interval
          in: query
          required: true
          description: Supported Binance Spot candle interval.
          schema:
            $ref: "#/components/schemas/CandleInterval"
        - name: before
          in: query
          required: false
          description: Return candles whose open time is before this cursor.
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          required: false
          description: Maximum number of candles to return.
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 200
      responses:
        "200":
          description: Candle page.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandlePageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/AccessDenied"
        "404":
          $ref: "#/components/responses/SymbolNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/instruments/{symbol}/grid-limits:
    get:
      operationId: getInstrumentGridLimits
      summary: Get the Binance spot grid price limits of an instrument
      description: |
        Returns what a Binance spot grid bot derives its price range limits
        from: the current 5-minute average price, the symbol's percent price
        multipliers (the stricter of PERCENT_PRICE_BY_SIDE and PERCENT_PRICE),
        and its PRICE_FILTER bounds and tick size. A value is 0 when the symbol
        has no such filter. The bot accepts a lower price down to
        `max(min_price, (1 − (1 − bid_multiplier_down) × 0.85) × average_price)`
        and an upper price up to
        `min(max_price, (1 + (ask_multiplier_up − 1) × 0.85) × average_price)`.
        Without a multiplier, clients allow 50% below or above the average
        price; a `max_price` of 0 sets no maximum. Grid prices are multiples of
        `tick_size`.
      tags: [Market Data]
      security:
        - sessionToken: []
      parameters:
        - $ref: "#/components/parameters/RequestID"
        - $ref: "#/components/parameters/Symbol"
      responses:
        "200":
          description: Grid limits.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GridLimitsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/AccessDenied"
        "404":
          $ref: "#/components/responses/SymbolNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/AnalysisUnavailable"
  /api/v1/chart/indicators:
    get:
      operationId: listChartIndicators
      summary: List the indicators charts of an interval request and how to draw them
      description: |
        Clients subscribe to live charts of the interval with each item's `type`
        and `parameters`, in this order, and draw the returned series as described
        here. The administrator configures the catalog, so it may be empty.
      tags: [Market Data]
      security: [{ sessionToken: [] }]
      parameters:
        - name: interval
          in: query
          required: true
          schema: { $ref: "#/components/schemas/CandleInterval" }
      responses:
        "200":
          description: Chart indicator catalog.
          content: { application/json: { schema: { $ref: "#/components/schemas/ChartIndicatorCatalog" } } }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/AccessDenied" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/favorites:
    get:
      operationId: listFavorites
      tags: [Favorites]
      security: [{ sessionToken: [] }]
      responses:
        "200":
          description: Current user's favorites.
          content: { application/json: { schema: { $ref: "#/components/schemas/FavoritesResponse" } } }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/AccessDenied" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/favorites/{symbol}:
    put:
      operationId: addFavorite
      tags: [Favorites]
      security: [{ sessionToken: [] }]
      parameters: [{ $ref: "#/components/parameters/Symbol" }]
      responses:
        "200":
          description: Favorite state.
          content: { application/json: { schema: { $ref: "#/components/schemas/Favorite" } } }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/AccessDenied" }
        "404": { $ref: "#/components/responses/SymbolNotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      operationId: removeFavorite
      tags: [Favorites]
      security: [{ sessionToken: [] }]
      parameters:
        - { $ref: "#/components/parameters/Symbol" }
        - name: confirm_alerts
          in: query
          required: false
          schema: { type: boolean, default: false }
      responses:
        "204": { description: "Favorite and, when confirmed, its alerts were removed." }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/AccessDenied" }
        "404": { $ref: "#/components/responses/FavoriteNotFound" }
        "409": { $ref: "#/components/responses/FavoriteConflict" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/favorites/analysis:
    post:
      operationId: analyzeFavorites
      summary: Analyze the current user's active favorite instruments
      tags: [Favorites]
      security: [{ sessionToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MarketAnalysisRequest" }
      responses:
        "200":
          description: Analysis results restricted to favorite instruments.
          content: { application/json: { schema: { $ref: "#/components/schemas/MarketAnalysisResponse" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/AccessDenied" }
        "422": { $ref: "#/components/responses/UnprocessableAnalysis" }
        "500": { $ref: "#/components/responses/InternalError" }
        "503": { $ref: "#/components/responses/AnalysisUnavailable" }
  /api/v1/instruments/{symbol}/alerts:
    get:
      operationId: listPriceAlerts
      tags: [Price Alerts]
      security: [{ sessionToken: [] }]
      parameters: [{ $ref: "#/components/parameters/Symbol" }]
      responses:
        "200":
          description: Current user's active alerts for the instrument.
          content: { application/json: { schema: { $ref: "#/components/schemas/PriceAlertsResponse" } } }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/AccessDenied" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      operationId: createPriceAlert
      tags: [Price Alerts]
      security: [{ sessionToken: [] }]
      parameters: [{ $ref: "#/components/parameters/Symbol" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/PriceAlertInput" } } }
      responses:
        "201":
          description: Alert created; the instrument is now a favorite.
          content: { application/json: { schema: { $ref: "#/components/schemas/PriceAlert" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/AccessDenied" }
        "404": { $ref: "#/components/responses/SymbolNotFound" }
        "409": { $ref: "#/components/responses/AlertConflict" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/alerts/{alert_id}:
    patch:
      operationId: updatePriceAlert
      tags: [Price Alerts]
      security: [{ sessionToken: [] }]
      parameters: [{ $ref: "#/components/parameters/AlertID" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/PriceAlertInput" } } }
      responses:
        "200":
          description: Updated alert.
          content: { application/json: { schema: { $ref: "#/components/schemas/PriceAlert" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/AccessDenied" }
        "404": { $ref: "#/components/responses/AlertNotFound" }
        "409": { $ref: "#/components/responses/AlertConflict" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      operationId: deletePriceAlert
      tags: [Price Alerts]
      security: [{ sessionToken: [] }]
      parameters: [{ $ref: "#/components/parameters/AlertID" }]
      responses:
        "204": { description: Alert deleted. }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/AccessDenied" }
        "404": { $ref: "#/components/responses/AlertNotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/me:
    get:
      operationId: getCurrentUser
      summary: Describe the authenticated user
      tags: [Administration]
      security: [{ sessionToken: [] }]
      responses:
        "200":
          description: Current user.
          content: { application/json: { schema: { $ref: "#/components/schemas/CurrentUser" } } }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/AccessDenied" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/admin/indicator-types:
    get:
      operationId: listIndicatorTypes
      summary: List every indicator the scanner can calculate
      description: Self-descriptions of the registered indicators, used to build the scanner indicator form.
      tags: [Administration]
      security: [{ sessionToken: [] }]
      responses:
        "200":
          description: Indicator types ordered by type.
          content: { application/json: { schema: { $ref: "#/components/schemas/IndicatorTypeCatalog" } } }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/AdministratorRequired" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/admin/scanner-indicators:
    get:
      operationId: listScannerIndicators
      summary: List the configured scanner indicators
      description: |
        Strategies can read every configured indicator. It is drawn on the
        charts of its interval when `show_in_chart` is set and shown as a market
        table column when `show_in_table` is set. Items are in display order, which orders the
        table columns and the chart indicators.
      tags: [Administration]
      security: [{ sessionToken: [] }]
      responses:
        "200":
          description: Configured indicators.
          content: { application/json: { schema: { $ref: "#/components/schemas/ScannerIndicatorList" } } }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/AdministratorRequired" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      operationId: createScannerIndicator
      summary: Add a scanner indicator on one or more intervals
      description: Adds one indicator per chosen interval, all or none, at the end of the display order.
      tags: [Administration]
      security: [{ sessionToken: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ScannerIndicatorInput" } } }
      responses:
        "201":
          description: Added indicators, in interval request order.
          content: { application/json: { schema: { $ref: "#/components/schemas/ScannerIndicatorList" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/AdministratorRequired" }
        "409": { $ref: "#/components/responses/ScannerIndicatorConflict" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      operationId: deleteUnusedScannerIndicators
      summary: Remove every scanner indicator no strategy reads
      description: |
        Indicators read by a strategy stay. Charts and tables no longer show the
        removed indicators, and their values are no longer tracked.
      tags: [Administration]
      security: [{ sessionToken: [] }]
      responses:
        "204": { description: Unused indicators removed. }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/AdministratorRequired" }
        "409": { $ref: "#/components/responses/ScannerIndicatorInUse" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/admin/scanner-indicator-batches:
    post:
      operationId: createScannerIndicatorBatch
      summary: Add several different scanner indicators
      description: |
        Adds the indicators, all or none, at the end of the display order,
        neither shown in tables nor drawn on charts. Used to add the missing
        indicators a strategy expression reads.
      tags: [Administration]
      security: [{ sessionToken: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ScannerIndicatorBatch" } } }
      responses:
        "201":
          description: Added indicators, in request order.
          content: { application/json: { schema: { $ref: "#/components/schemas/ScannerIndicatorList" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/AdministratorRequired" }
        "409": { $ref: "#/components/responses/ScannerIndicatorConflict" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/admin/scanner-indicator-order:
    put:
      operationId: reorderScannerIndicators
      summary: Set the display order of the scanner indicators
      tags: [Administration]
      security: [{ sessionToken: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ScannerIndicatorOrder" } } }
      responses:
        "200":
          description: Indicators in the new order.
          content: { application/json: { schema: { $ref: "#/components/schemas/ScannerIndicatorList" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/AdministratorRequired" }
        "404": { $ref: "#/components/responses/ScannerIndicatorNotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/admin/scanner-indicators/{indicator_id}:
    patch:
      operationId: updateScannerIndicator
      summary: Change the table visibility and pane scale of a scanner indicator
      tags: [Administration]
      security: [{ sessionToken: [] }]
      parameters: [{ $ref: "#/components/parameters/ScannerIndicatorID" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ScannerIndicatorUpdate" } } }
      responses:
        "200":
          description: Updated indicator.
          content: { application/json: { schema: { $ref: "#/components/schemas/ScannerIndicator" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/AdministratorRequired" }
        "404": { $ref: "#/components/responses/ScannerIndicatorNotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      operationId: deleteScannerIndicator
      summary: Remove a scanner indicator
      description: Rejected while a strategy reads the indicator.
      tags: [Administration]
      security: [{ sessionToken: [] }]
      parameters: [{ $ref: "#/components/parameters/ScannerIndicatorID" }]
      responses:
        "204": { description: Indicator removed. }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/AdministratorRequired" }
        "404": { $ref: "#/components/responses/ScannerIndicatorNotFound" }
        "409": { $ref: "#/components/responses/ScannerIndicatorInUse" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/admin/users:
    get:
      operationId: listUsers
      summary: List the application users
      description: Users are added through the Telegram bot. The administrator comes first.
      tags: [Administration]
      security: [{ sessionToken: [] }]
      responses:
        "200":
          description: Application users.
          content: { application/json: { schema: { $ref: "#/components/schemas/UserList" } } }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/AdministratorRequired" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    telegramInitData:
      type: apiKey
      in: header
      name: Authorization
      description: Telegram Mini App init data prefixed with `tma `; accepted only by `createSession`.
    sessionToken:
      type: http
      scheme: bearer
      description: Session token issued by `createSession`.
  parameters:
    RequestID:
      name: X-Request-ID
      in: header
      required: false
      description: Optional caller-provided correlation ID. Unsafe values are replaced.
      schema:
        type: string
    Symbol:
      name: symbol
      in: path
      required: true
      description: Market symbol.
      schema:
        type: string
        minLength: 1
        maxLength: 32
...
  headers:
    RequestID:
      description: Correlation ID assigned to the request.
      required: true
      schema:
        type: string
...
  schemas:
    Session:
      type: object
      additionalProperties: false
      required: [token, expires_at]
      properties:
        token:
          type: string
          description: Opaque bearer token; it also expires after a period without use.
        expires_at:
          type: string
          format: date-time
          description: Time after which the session ends regardless of use.
...